Fixed R yaml.load_file
It was bugging the hell out of me at the meetup. You were right to have as.data.frame passed to yaml.load_file. This returns a list of data frames. The only thing we were missing is to have all of the data frames unlisted and bound by rows ("rebind", yaml_data).